configure: don't locate gapi's exe files, just the scripts
If the scripts are not in the path, there are very low chances
that the .exe files would be, instead (and using `which` instead
of AC_PATH_PROG for them doesn't make much sense; plus it was
buggy [1]). This should have probably been a search with `find`,
but it's better to just remove it altogether, and just leave the
tests for the wrapper scripts, which should be found anyway.

[1] https://github.com/meebey/messagingmenu-sharp/commit/68bebcd0a499a7640f9063d67459868b21e5c02c